% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lollipop.R, R/geom_post.R
\name{geom_lollipop}
\alias{geom_lollipop}
\title{Title}
\usage{
geom_lollipop()

geom_lollipop()
}
\value{



}
\description{
Title

Title
}
\examples{
ggplot(cars) +
aes(x = speed, y = dist) +
geom_lollipop()
ggplot(cars) +
aes(x = speed, y = dist) +
geom_lollipop()
}
