% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_post.R
\name{geom_post}
\alias{geom_post}
\title{Post geom}
\usage{
geom_post(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
a ggplot2 geom layer
}
\description{
geom_post() may be used to visualize discrete quantitative distributions
}
\examples{
library(ggplot2)
ggplot(data = prize_wheel) +
aes(x = payout, y = frequency) +
geom_post()
}
