% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_post_label.R
\name{geom_post_label}
\alias{geom_post_label}
\title{Post geom}
\usage{
geom_post_label(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  vjust = -0.6,
  ...
)
}
\value{
a ggplot2 geom layer
}
\description{
geom_post() may be used to visualize discrete quantitative distributions
}
\examples{
library(ggplot2)
ggplot(data = prize_wheel) +
aes(x = payout, y = frequency) +
geom_post() +
geom_post_label()

tidy_dbinom(single_trial_prob = 1/6, num_trials = 10) \%>\%   # from ma206distributions
ggplot() +
 aes(x = num_successes,
     y = probability) +
 geom_lollipop(annotate = TRUE,
               round_digits = 2) +                            # from ma206distributions
 scale_x_counting()                 # from ma206distributions

}
