% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prize_wheel}
\alias{prize_wheel}
\title{Example problem data}
\format{
\subsection{\code{prize_wheel}}{

A data frame with 3 rows and 3 columns:
\describe{
\item{sector_type}{type of pie slice from prize wheel}
\item{frequency}{proportion of pie slice of that type}
\item{payout}{how much is paid out in the case landing on slice}
...
}
}
}
\source{
\if{html}{\out{
<ISI chapter II>
}}
}
\usage{
prize_wheel
}
\description{
Data from chapter 11 of MA206's text book
}
\examples{
prize_wheel

library(ggplot2)
ggplot(prize_wheel) +
    aes(x = payout, y = frequency) +
    geom_lollipop()

}
\keyword{datasets}
