% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_counting.R
\name{scale_x_counting}
\alias{scale_x_counting}
\title{scale_x_counting}
\usage{
scale_x_counting(breaks = 0:10000, ...)
}
\arguments{
\item{...}{}
}
\description{
Show x axis guides with ticks at integers
}
\examples{
library(ggplot2)
ggplot(tidy_dbinom(1/6, num_trials = 8)) +
  aes(x = num_successes) +
  scale_x_counting() +
  aes(y = probability) +
  geom_lollipop()

}
