% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_dbinom.R
\name{stamp_dbinom}
\alias{stamp_dbinom}
\title{stamp_dbinom}
\usage{
stamp_dbinom(single_trial_prob = 0.5, num_trials = 10, annotate = FALSE)
}
\arguments{
\item{single_trial_prob}{numeric value between 0 and 1}

\item{num_trials}{positive integer which is the total number of trials}
}
\value{
A plot that plots the number of success with associated probabilities
}
\description{
stamp_dbinom returns a plot that plots the number of successes with probabilities given the probability of a single trial and the total number of trials.
}
\examples{
library(ggplot2)
ggplot() +
stamp_dbinom()

ggplot() +
stamp_dbinom(single_trial_prob = 1/12,
num_trials = 3,
annotate=TRUE)

}
