% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_dgeometric.R
\name{stamp_dgeometric}
\alias{stamp_dgeometric}
\title{Title}
\usage{
stamp_dgeometric(single_prob = 0.5, num_attempts = 10)
}
\arguments{
\item{num_attempts}{}
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot() +
stamp_dgeometric()
}
