% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_counting.R
\name{theme_axis_x_truncated}
\alias{theme_axis_x_truncated}
\title{theme_axis_x_truncated}
\usage{
theme_axis_x_truncated()
}
\arguments{
\item{breaks}{numeric vector where breaks should be placed}

\item{...}{}
}
\value{
ggplot2 theme function to indicate that x is right censored
}
\description{
Show x axis guides with ticks at integers
}
\examples{
library(ggplot2)
ggplot(tidy_dbinom(1/6, num_trials = 8)) +
  aes(x = num_successes) +
  scale_x_counting() +
  aes(y = probability) +
  geom_lollipop() +
  theme_axis_x_truncated()
}
