% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dbinom.R
\name{tidy_dbinom}
\alias{tidy_dbinom}
\title{Tidy_dbinom returns a data table that pairs number of successes with probabilities given the probability of a single trial and the total number of trials.}
\usage{
tidy_dbinom(single_trial_prob = 0.5, num_trials = 10)
}
\arguments{
\item{single_trial_prob}{a numeric value between 0 and 1}

\item{num_trials}{positive integer which is the total number of trials}
}
\value{
A data frame pairing number of successes with probabilities.
}
\description{
Tidy_dbinom returns a data table that pairs number of successes with probabilities given the probability of a single trial and the total number of trials.
}
\examples{
tidy_dbinom(single_trial_prob = 1/6, num_trials = 4)
tidy_dbinom(.5, 10)
tidy_dbinom(1/12, 3)
tidy_dbinom(.5, 20)

library(ggplot2)
ggplot(tidy_dbinom(1/6, num_trials = 8)) +
  aes(x = num_successes) +
  aes(y = probability) +
  geom_lollipop()

}
