% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dgeometric.R
\name{tidy_dgeometric}
\alias{tidy_dgeometric}
\title{Title}
\usage{
tidy_dgeometric(single_trial_prob = 0.5, num_attempts = 5)
}
\arguments{
\item{num_attempts}{}
}
\description{
Title
}
\examples{
tidy_dgeometric()
tidy_dgeometric(single_trial_prob = 1/6, num_attempts = 10)
}
