% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_trials.R
\name{cross_trials}
\alias{cross_trials}
\title{Title}
\usage{
cross_trials(trial = prize_wheel, num_trials = 2)
}
\arguments{
\item{num_trials}{}
}
\description{
Title
}
\examples{
library(dplyr)
prize_wheel |>
   mutate(prob = frequency/sum(frequency)) |>
   cross_trials(num_trials = 2)

cross_trials(bernoulli_trial(.1), num_trials = 3)

library(ggplot2)
cross_trials(prize_wheel, num_trials = 2) |>
ggplot() +
aes(x = t1_sector_type, y = t2_sector_type)
}
