% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fulcrum.R
\name{geom_fulcrum}
\alias{geom_fulcrum}
\title{Title}
\usage{
geom_fulcrum(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(x = speed, y = dist) +
geom_point() +
geom_fulcrum(color = "red", shape = 2, size = 5)
}
