% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lollipop.R
\name{geom_lollipop}
\alias{geom_lollipop}
\title{Lollipop geom}
\usage{
geom_lollipop(...)
}
\value{
a ggplot geom layer
}
\description{
The lollipop is often used to attractively communicate about discrete quantitative distributions.
}
\examples{
library(ggplot2)
ggplot(data = prize_wheel) +
aes(x = payout, y = frequency) +
geom_lollipop()
}
