% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_post.R
\name{geom_post}
\alias{geom_post}
\title{Title}
\usage{
geom_post(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(x = speed, y = dist) +
geom_post()
}
